refactor: Cursor 로직 분리에 따라 Draft도 따로 분리
